requirements.txt
torch>=2.1.0
torchvision>=0.16.0
streamlit>=1.30
pillow>=10.0
transformers>=4.40
skorch>=1.3
pandas>=2.0
numpy>=1.26
seaborn>=0.13
matplotlib>=3.8
tqdm>=4.66
scikit-learn>=1.4
scipy>=1.11
ultralytics>=8.4